% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argosData-package.R
\docType{package}
\name{argosData-package}
\alias{argosData}
\alias{argosData-package}
\title{argosData: Data Package for Bearded Seal (Erignathus barbatus) Satellite Data}
\description{
This is a data package that contains made up satellite data for bearded 
    seals captured and tagged in Alaskan waters - ebArgos. The data includes the seal id, 
    locations (latitude and longitude), date and time of locations, and Argos 
    location quality.
}
\author{
\strong{Maintainer}: Rowenna Gryba \email{r.gryba@stat.ubc.ca}

}
\keyword{internal}
