% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebArgos.R
\docType{data}
\name{ebArgos}
\alias{ebArgos}
\title{Argos data for bearded seals tagged in Alaskan waters}
\format{Character, integer, factor}
\source{
Project conducted as a part of the BOEM project MARES.
}
\usage{
ebArgos
}
\description{
??????
}
\keyword{datasets}
