% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebDoc.R
\docType{data}
\name{ebArgos}
\alias{ebArgos}
\title{Argos satellite data for bearded seals (Erignathus barbatus)}
\format{The data format is a combination if integers, factors and date.}
\source{
The data is a product of Argos satellite tagging in Alaskan waters
}
\usage{
ebArgos
}
\description{
??? Fill in
}
\keyword{datasets}
